+++ modules/Users/Exceptions.php
@@ -208,8 +208,8 @@ if ( UserStaffID()
 			echo '<table class="widefat fixed-col valign-middle"><tr><th class="align-right"><label>' . _( 'Can Use' ) . ' ' .
 				( AllowEdit() ?
 				'<input type="checkbox" name="' . AttrEscape( 'can_use_' . $modcat ) .
-				'" onclick="' . AttrEscape( 'checkAll(this.form,this.form.can_use_' . $modcat .
-					'.checked,"can_use[' . $modcat . '");' ) . '" /> ' :
+				// @since 12.5 CSP remove unsafe-inline Javascript
+				'" class="onclick-checkall" data-name-like="' . AttrEscape( 'can_use[' . $modcat ) . '" /> ' :
 				'' ) .
 				'</label></th>';
 
@@ -222,8 +222,8 @@ if ( UserStaffID()
 				echo '<th class="align-right"><label>' . _( 'Can Edit' ) . ' ' .
 					( AllowEdit() ?
 					'<input type="checkbox" name="' . AttrEscape( 'can_edit_' . $modcat ) .
-					'" onclick="' . AttrEscape( 'checkAll(this.form,this.form.can_edit_' . $modcat .
-						'.checked,"can_edit[' . $modcat . '");' ) . '" /> ' :
+					// @since 12.5 CSP remove unsafe-inline Javascript
+					'" class="onclick-checkall" data-name-like="' . AttrEscape( 'can_edit[' . $modcat ) . '" /> ' :
 					'' ) .
 					'</label></th>';
 			}